MODCLUSTER-575 Move org.jboss.modcluster.load.SimpleLoadBalanceFactorProviderFactory to the impl package to avoid jar package conflict
